SharedInteractionSystem
Created SharedInteractionSystem and inherrited the Server-side InteractionSystem from it.

Anything in SharedInteractionSystem will be accessible from the client space, be cautious.
